+++ src/app/configs/config.ts
@@ -1,9 +1,12 @@
 import { WorkspaceConfiguration } from 'vscode';
 
 import {
+  ACTIVATE_MENU,
   AUTO_IMPORT,
+  CUSTOM_COMMANDS,
   EXCLUDE,
   INCLUDE,
+  MenuInterface,
   ORM,
   SHOW_PATH,
   WATCH,
@@ -21,6 +24,8 @@ import {
  * @property {string[]} exclude - The files to exclude
  * @property {string[]} watch - The files to watch
  * @property {boolean} showPath - Whether to show the path or not
+ * @property {object[]} customCommands - The custom commands
+ * @property {object} activateItem - Whether to show the menu or not
  * @property {boolean} autoImport - The auto import setting
  * @property {string} orm - The orm
  * @example
@@ -77,6 +82,30 @@ export class Config {
    */
   showPath: boolean;
   /**
+   * The custom commands.
+   * @type {object[]}
+   * @public
+   * @memberof Config
+   * @example
+   * const config = new Config(workspace.getConfiguration());
+   * console.log(config.customCommands);
+   * console.log(config.customCommands[0].name);
+   * console.log(config.customCommands[0].command);
+   * console.log(config.customCommands[0].args);
+   */
+  customCommands: object[];
+  /**
+   * Whether to show the menu or not.
+   * @type {MenuInterface}
+   * @public
+   * @memberof Config
+   * @example
+   * const config = new Config(workspace.getConfiguration());
+   * console.log(config.activateItem);
+   * console.log(config.activateItem.terminal.components);
+   */
+  activateItem: MenuInterface;
+  /**
    * The auto import setting.
    * @type {boolean}
    * @public
@@ -114,6 +143,10 @@ export class Config {
     this.exclude = config.get<string[]>('files.exclude') ?? EXCLUDE;
     this.watch = config.get<string[]>('files.watch') ?? WATCH;
     this.showPath = config.get<boolean>('files.showPath') ?? SHOW_PATH;
+    this.customCommands =
+      config.get<object[]>('submenu.customCommands') ?? CUSTOM_COMMANDS;
+    this.activateItem =
+      config.get<MenuInterface>('submenu.activateItem') ?? ACTIVATE_MENU;
     this.autoImport = config.get<boolean>('files.autoImport') ?? AUTO_IMPORT;
     this.orm = config.get<string>('files.orm') ?? ORM;
   }